feat: ✨ useStockPrice로 추상화
TextProps<"span"> 적용
