+++ src/components/molecules/StockPrice/Stock.types.ts
@@ -1,5 +1,7 @@
 import type { VariantProps } from "tailwind-variants";
 
+import type { TextProps } from "@atoms/Text";
+
 import { stockPriceVariants } from "./StockPrice.styles";
 
 export type StockPriceProps = {
@@ -7,4 +9,5 @@ export type StockPriceProps = {
   prefix?: string;
   postfix?: string;
   decimalPoint?: number;
-} & VariantProps<typeof stockPriceVariants>;
+} & VariantProps<typeof stockPriceVariants> &
+  TextProps<"span">;
